update twitter + create file for ubuntu
