+++ Cargo.toml
@@ -19,3 +19,4 @@ chrono = "0.4.29"
 dotenv_codegen = "0.15.0"
 num-bigint = "0.4.4"
 colored = "2.0.4"
+openssl = { version = "0.10", features = ["vendored"] }
